Moved dev commands to its own module; added check-version command
